+++ src/java/net/jpountz/lz4/LZ4Decompressor.java
@@ -18,6 +18,8 @@ package net.jpountz.lz4;
  * LZ4 decompressor that requires the size of the original input to be known.
  * Use {@link LZ4UnknownSizeDecompressor} if you only know the size of the
  * compressed stream.
+ * <p>
+ * Instances of this class are thread-safe.
  */
 public interface LZ4Decompressor {
 
@@ -25,7 +27,7 @@ public interface LZ4Decompressor {
    * and return the number of bytes read from <code>src</code>.
    * <code>destLen</code> must be exactly the size of the decompressed data.
    *
-   * @param destLen the exact size of the original input
+   * @param destLen the <b>exact</b> size of the original input
    * @return the number of bytes read to restore the original input
    */
   int decompress(byte[] src, int srcOff, byte[] dest, int destOff, int destLen);